Restructured tests
